app/core/dependencies.py
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from jose import JWTError, jwt
from app.core.config import settings
from app.core.jwt import decode_token
from app.common.exceptions import UnauthorizedError, ForbiddenError
from app.modules.users.schemas import User

security = HTTPBearer()

async def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(security)) -> User:
    token = credentials.credentials
    payload = decode_token(token)
    if payload is None:
        raise UnauthorizedError("Could not validate credentials")
    
    user_id: str = payload.get("sub")
    phone: str = payload.get("phone")
    apps: list = payload.get("apps")
    
    if user_id is None or phone is None:
        raise UnauthorizedError("Invalid token payload")
        
    # In a real app, we might want to fetch the user from DB to ensure they still exist/are active.
    # For now, we trust the token as per "Stateless" JWT, but the prompt says "Single Identity Model".
    # Let's construct the User object from token data to avoid DB hit on every request if possible,
    # OR fetch from DB. Fetching from DB is safer.
    
    # For this implementation, I'll construct a basic User object from the token to be fast,
    # but ideally we should check the DB.
    
    return User(
        id=user_id,
        phone_number=phone,
        apps_enabled=apps,
        created_at=datetime.fromtimestamp(payload.get("iat", 0)) # Approximate
    )

from datetime import datetime

def get_current_active_user(current_user: User = Depends(get_current_user)) -> User:
    # Add logic to check if user is active if needed
    return current_user

def require_app_permission(app_name: str):
    def _require_app_permission(current_user: User = Depends(get_current_user)):
        if app_name not in current_user.apps_enabled:
            raise ForbiddenError(f"User does not have access to {app_name}")
        return current_user
    return _require_app_permission
